size_t istället för int på rimloiga ställen, dokumnetation uppdaterad

+++ linked_list.h
@@ -40,7 +40,7 @@ int ioopm_list_last(ioopm_list_t *list);
 /// @param list the linked list that will be extended
 /// @param index the position in the list
 /// @param value the value to be inserted
-void ioopm_list_insert(ioopm_list_t *list, int index, int value);
+void ioopm_list_insert(ioopm_list_t *list, size_t index, int value);
 
 /// @brief Remove an element from a linked list in O(n) time.
 /// The valid values of index are [0,n-1] for a list of n elements,
@@ -48,8 +48,9 @@ void ioopm_list_insert(ioopm_list_t *list, int index, int value);
 /// @pre 0 <= index < length(list)
 /// @param list the linked list
 /// @param index the position in the list
-/// @return the value removed
-bool ioopm_list_remove(ioopm_list_t *list, int index, int *result);
+/// @return true if the value was found and removed
+/// A side effect is that the removed value gets put in the addres of result
+bool ioopm_list_remove(ioopm_list_t *list, size_t index, int *result);
 
 /// @brief Retrieve an element from a linked list in O(n) time.
 /// The valid values of index are [0,n-1] for a list of n elements,
@@ -57,13 +58,13 @@ bool ioopm_list_remove(ioopm_list_t *list, int index, int *result);
 /// @pre 0 <= index < length(list)
 /// @param list the linked list that will be extended   
 /// @param index the position in the list
-/// @return the value at the given position
-bool ioopm_list_get(ioopm_list_t *list, int index, int *result);
+/// @return true if the value was found at the index, side effect is that the value gets put in the address of result
+bool ioopm_list_get(ioopm_list_t *list, size_t index, int *result);
 
 /// @brief Lookup the number of elements in the linked list in O(1) time
 /// @param list the linked list
 /// @return the number of elements in the list
-int ioopm_list_size(ioopm_list_t *list);
+size_t ioopm_list_size(ioopm_list_t *list);
 
 /// @brief Test whether a list is empty or not
 /// @param list the linked list
